gitserver: replace deprecated P4Exec command with stub unimplemented interface (#59999)
As a follow up to https://github.com/sourcegraph/sourcegraph/pull/59997, this PR replaces the implementation of the `P4Exec` RPC with a stub that simply returns `codes.Unimplemented` (see https://grpc.github.io/grpc/core/md_doc_statuscodes.html for more detail on grpc status codes). This avoids the need to refactor this usage of `p4` to use the new `perforce.NewBaseCommand` abstraction. 

This RPC endpoint is already unused by our current gitserver client. 

Any old gitserver clients that still contact this endpoint (e.g during an upgrade) will receive the Unimplemented error. However, this error shouldn't prevent a rollout - so this brief temporary state should be fine. 


## Test plan

CI
