+++ cmd/gitserver/internal/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
         "gitservice.go",
         "list_gitolite.go",
         "lock.go",
-        "p4exec.go",
         "patch.go",
         "repo_info.go",
         "search.go",
@@ -95,7 +94,6 @@ go_test(
         "list_gitolite_test.go",
         "main_test.go",
         "mocks_test.go",
-        "p4exec_test.go",
         "repo_info_test.go",
         "server_grpc_test.go",
         "server_test.go",
