+++ cmd/gitserver/internal/server.go
@@ -836,6 +836,15 @@ func (s *Server) CloneRepo(ctx context.Context, repo api.RepoName, opts CloneOpt
 	return "", nil
 }
 
+func (s *Server) P4Exec(_ context.Context, logger log.Logger, _ *p4ExecRequest, _ io.Writer) execStatus {
+	logger.Error("p4exec has been deprecated and removed on gitserver")
+
+	return execStatus{
+		ExitStatus: 1,
+		Err:        errors.New("p4exec has been deprecated and removed on gitserver"),
+	}
+}
+
 func (s *Server) doClone(
 	ctx context.Context,
 	repo api.RepoName,
